requirements.txt
Flask==3.1.0
Flask_Login==0.6.3
marshmallow==3.26.1
psycopg2_binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
